% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrs.R
\name{api_adr_pzn}
\alias{api_adr_pzn}
\title{Get ADRs for a list of PZNs}
\usage{
api_adr_pzn(
  creds,
  pzns,
  lang = c("english", "german", "german-simple"),
  details = FALSE
)
}
\arguments{
\item{creds}{A list containing the access token and host}

\item{pzns}{A vector with the PZNs}

\item{lang}{A string with the language. One of \code{english}, \code{german} or \code{german-simple}}
}
\value{
A data frame with the ADRs
}
\description{
Get ADRs for a list of PZNs
}
\examples{
\dontrun{
creds <- api_login("https://api.example.com", "username", "password", "user")
api_adr_pzn(creds, c("PZN1", "PZN2"), lang = "english")
}
}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.
}
