% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{api_compound_name}
\alias{api_compound_name}
\title{Get names for a list of compound names}
\usage{
api_compound_name(creds, compounds)
}
\arguments{
\item{creds}{A list containing the access token and host}

\item{compounds}{A vector with the compound names}
}
\value{
A list of data frames with the interactions
}
\description{
Get names for a list of compound names
}
\examples{
\dontrun{
creds <- api_login("https://api.example.com", "username", "password", "user")
api_compound_name(creds, compounds = c("Aspirin", "Paracetamol"))
}
}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.
}
