% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulation.R
\name{api_formulations}
\alias{api_formulations}
\title{Get formulations}
\usage{
api_formulations(creds)
}
\arguments{
\item{creds}{A list containing the access token and host}
}
\value{
A data frame with the formulations
}
\description{
Get formulations
}
\examples{
\dontrun{
creds <- api_login("https://api.example.com", "username", "password", "user")
api_formulations(creds)
}
}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.
}
