% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{api_get_users}
\alias{api_get_users}
\title{List users}
\usage{
api_get_users(creds)
}
\arguments{
\item{creds}{A list with the access token and the host.}
}
\value{
A message with the result of the operation.
}
\description{
This function lists all users in the system. Admin only.
}
\examples{
\dontrun{
creds <- api_login("https://api.example.com", "user", "password", "user")
api_get_users(creds)
}
}
