% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.R
\name{api_info}
\alias{api_info}
\title{Retrieve API Information}
\usage{
api_info(x)
}
\arguments{
\item{x}{A character string specifying the API host for which information is to be retrieved or
an object of class \link{ApiCredentials}.}
}
\value{
A list containing information about the specified API. Throws an error if the API is not available.
}
\description{
This function retrieves information about a specified API.
}
\examples{
\dontrun{
api_info("https://api.example.com")

creds <- api_login("https://api.example.com", "username", "password")
api_info(creds)
}

}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.

\code{\link[=api_ping]{api_ping()}} to check if the API is available.
}
