% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{api_interaction_compound_batch}
\alias{api_interaction_compound_batch}
\title{Get interactions for a list of compound batches}
\usage{
api_interaction_compound_batch(creds, compound_batches)
}
\arguments{
\item{creds}{A list containing the access token and host}

\item{compound_batches}{A list of lists containing compounds, id, a \code{details} boolean, and a \code{doses} boolean}
}
\value{
A list of data frames with the interactions
}
\description{
Get interactions for a list of compound batches
}
\examples{
\dontrun{
creds <- api_login("https://api.example.com", "username", "password", "user")
api_interaction_compound_batch(creds, list(list(
  compounds = c("Aspirin", "Paracetamol"),
  id = 1, details = TRUE, doses = FALSE
)))
}
}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.
}
