% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{api_interaction_pzn}
\alias{api_interaction_pzn}
\title{Get interactions for a list of PZNs}
\usage{
api_interaction_pzn(creds, pzns, details = FALSE)
}
\arguments{
\item{creds}{A list containing the access token and host}

\item{pzns}{A vector with the PZNs}

\item{details}{A boolean indicating if detailed information should be returned}
}
\value{
A data frame with the interactions
}
\description{
Get interactions for a list of PZNs
}
\examples{
\dontrun{
creds <- api_login("https://api.example.com", "username", "password", "user")
api_interaction_pzn(creds, c("PZN1", "PZN2"), details = TRUE)
}
}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.
}
