% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{api_interaction_pzn_batch}
\alias{api_interaction_pzn_batch}
\title{Get interactions for a list of PZN batches}
\usage{
api_interaction_pzn_batch(creds, pzn_batches)
}
\arguments{
\item{creds}{A list containing the access token and host}

\item{pzn_batches}{A list of lists containing PZNs, id and a \code{details} boolean}
}
\value{
A list of data frames with the interactions
}
\description{
Get interactions for a list of PZN batches
}
\examples{
\dontrun{
creds <- api_login("https://api.example.com", "username", "password", "user")
api_interaction_pzn_batch(creds, list(list(pzns = c("PZN1", "PZN2"), id = 1, details = TRUE)))
}
}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.
}
