% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{api_login}
\alias{api_login}
\title{Login to the API}
\usage{
api_login(host, user, password, role = c("default", "user", "admin"))
}
\arguments{
\item{host}{The host of the API}

\item{user}{The username}

\item{password}{The password}

\item{role}{The role of the user. Any of "user" or "admin".}
}
\value{
An \code{ApiCredentials} object with the login information.
}
\description{
Login to the API
}
\examples{
\dontrun{
api_login("https://api.example.com", "user", "password", "user")
}
}
\seealso{
\link{ApiCredentials} for the object structure.

\code{\link[=api_user_init_password]{api_user_init_password()}} to set the password.
}
