% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.R
\name{api_ping}
\alias{api_ping}
\title{Ping the API}
\usage{
api_ping(x)
}
\arguments{
\item{x}{A character string specifying the API host to be pinged or an object of class \link{ApiCredentials}.}
}
\value{
A list containing the response from the API. Throws an error if the API is not available.
}
\description{
This function pings the  API to check if it is available.
}
\examples{
\dontrun{
api_ping("https://api.example.com")

creds <- api_login("https://api.example.com", "username", "password")
api_ping(creds)
}
}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.

\code{\link[=api_info]{api_info()}} to retrieve information about the API.
}
