% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priscus.R
\name{api_priscus_pzn}
\alias{api_priscus_pzn}
\title{Get PRISCUS status for a list of PZNs}
\usage{
api_priscus_pzn(creds, pzns)
}
\arguments{
\item{creds}{A list containing the access token and host}

\item{pzns}{A vector with the PZNs}
}
\value{
A data frame with the priscus status
}
\description{
Get PRISCUS status for a list of PZNs
}
\examples{
\dontrun{
creds <- api_login("https://api.example.com", "username", "password", "user")
api_priscus_pzn(creds, c("PZN1", "PZN2"))
}
}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.
}
