% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_drugs.R
\name{api_qt_compound}
\alias{api_qt_compound}
\title{Get QT category for a list of compound names}
\usage{
api_qt_compound(creds, compounds)
}
\arguments{
\item{creds}{A list containing the access token and host}

\item{compounds}{A vector with compound names}
}
\value{
A data frame with the QT category
}
\description{
Get QT category for a list of compound names
}
\examples{
\dontrun{
creds <- api_login("https://api.example.com", "username", "password", "user")
api_qt_compound(creds, c("Metoprolol", "Aspirin"))
}
}
\seealso{
\code{\link[=api_login]{api_login()}} to retrieve the login object.
}
