% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{api_user_init_password}
\alias{api_user_init_password}
\title{Set password}
\usage{
api_user_init_password(host, user, token, password)
}
\arguments{
\item{host}{The host of the API}

\item{user}{The username/email}

\item{token}{The token received from the sign-up email}

\item{password}{The new password}
}
\value{
A list with the response and the host
}
\description{
Set password
}
\examples{
\dontrun{
api_user_init_password("https://api.example.com", "test@user.com", "token", "password")
}
}
