% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{api_login}
\alias{api_login}
\title{Login to the API}
\usage{
api_login(host, user, password, role = NULL)
}
\arguments{
\item{host}{The host of the API}

\item{user}{The username}

\item{password}{The password}

\item{role}{The role of the user. Any of "user" or "admin".}
}
\value{
A list with the access token and the host
}
\description{
Login to the API
}
\examples{
\dontrun{
api_login("https://api.example.com", "user", "password", "user")
}
}
